+++ plugins/input/csv/csv_index_featureset.hpp
@@ -40,7 +40,7 @@
 
 class csv_index_featureset : public mapnik::Featureset
 {
-    using value_type = std::pair<std::size_t, std::size_t>;
+    using value_type = std::pair<std::uint64_t, std::uint64_t>;
     using locator_type = csv_utils::geometry_column_locator;
 public:
